Use a view controller for the timeline.
This exposes lifecycle methods and tidies up the Coordinator.
