+++ ElementX/Sources/Screens/RoomScreen/RoomScreenModels.swift
@@ -63,12 +63,6 @@ struct RoomScreenViewState: BindableState {
     }
     
     let scrollToBottomPublisher = PassthroughSubject<Void, Never>()
-    
-    /// Returns the opacity that the supplied timeline item's cell should be.
-    func opacity(for item: RoomTimelineViewProvider) -> CGFloat {
-        guard case let .reply(selectedItemID, _) = composerMode else { return 1.0 }
-        return selectedItemID == item.id ? 1.0 : 0.5
-    }
 }
 
 struct RoomScreenViewStateBindings {
